# docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "这是一个使用 Golang (Gin, GORM, Wire) 构建的基础系统 API 服务第二版。",
        "title": "Go Base System API V2",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "https://www.example.com/support",
            "email": "support@example.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1 // **确保与 NewGinEngine 中注册的路由组路径一致**",
    "paths": {
        "/health": {
            "get": {
                "description": "检查API服务是否正常运行",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "健康检查"
                ],
                "summary": "服务健康检查",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/me/profile": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "获取当前通过 JWT 认证的用户的详细信息。",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "用户 (User)"
                ],
                "summary": "获取当前登录用户信息 (Get Current User Profile)",
                "responses": {
                    "200": {
                        "description": "成功获取当前用户信息",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.ResponseData"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/model.User"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "未授权或 Token 无效",
                        "schema": {
                            "$ref": "#/definitions/response.ResponseData"
                        }
                    },
                    "404": {
                        "description": "用户未找到 (例如，Token 中的用户 ID 在数据库中不存在)",
                        "schema": {
                            "$ref": "#/definitions/response.ResponseData"
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "$ref": "#/definitions/response.ResponseData"
                        }
                    }
                }
            }
        },
        "/users/id/{id}": {
            "get": {
                "description": "通过指定的用户ID检索用户的详细信息。",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "用户 (User)"
                ],
                "summary": "根据用户ID获取用户信息 (Get User By ID)",
                "parameters": [
                    {
                        "type": "integer",
                        "format": "uint",
                        "description": "用户ID (User ID)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功获取用户信息",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.ResponseData"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/model.User"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "无效的用户ID",
                        "schema": {
                            "$ref": "#/definitions/response.ResponseData"
                        }
                    },
                    "404": {
                        "description": "用户未找到",
                        "schema": {
                            "$ref": "#/definitions/response.ResponseData"
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "$ref": "#/definitions/response.ResponseData"
                        }
                    }
                }
            }
        },
        "/users/login": {
            "post": {
                "description": "使用用户名/邮箱和密码进行登录，成功后返回 JWT。",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "用户认证 (User Authentication)"
                ],
                "summary": "用户登录 (User Login)",
                "parameters": [
                    {
                        "description": "用户登录请求体",
                        "name": "user_login_req",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.UserLoginReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "登录成功，返回 JWT",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.ResponseData"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/dto.UserLoginRes"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "请求参数错误",
                        "schema": {
                            "$ref": "#/definitions/response.ResponseData"
                        }
                    },
                    "401": {
                        "description": "认证失败 (用户名或密码错误)",
                        "schema": {
                            "$ref": "#/definitions/response.ResponseData"
                        }
                    },
                    "500": {
                        "description": "服务器内部错误 (例如，Token 生成失败)",
                        "schema": {
                            "$ref": "#/definitions/response.ResponseData"
                        }
                    }
                }
            }
        },
        "/users/register": {
            "post": {
                "description": "根据提供的用户名、邮箱和密码创建一个新用户。",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "用户认证 (User Authentication) // 将注册和登录归为同一 Tag"
                ],
                "summary": "用户注册 (Register User)",
                "parameters": [
                    {
                        "description": "用户注册请求体",
                        "name": "user_register_req",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.UserRegisterReq"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "用户创建成功",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.ResponseData"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/model.User"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "请求参数错误",
                        "schema": {
                            "$ref": "#/definitions/response.ResponseData"
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "$ref": "#/definitions/response.ResponseData"
                        }
                    }
                }
            }
        },
        "/users/username/{username}": {
            "get": {
                "description": "通过指定的用户名检索用户的详细信息。",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "用户 (User)"
                ],
                "summary": "根据用户名获取用户信息 (Get User By Username)",
                "parameters": [
                    {
                        "maxLength": 50,
                        "minLength": 3,
                        "type": "string",
                        "description": "用户名 (Username)",
                        "name": "username",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功获取用户信息",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.ResponseData"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/model.User"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "请求参数错误",
                        "schema": {
                            "$ref": "#/definitions/response.ResponseData"
                        }
                    },
                    "404": {
                        "description": "用户未找到",
                        "schema": {
                            "$ref": "#/definitions/response.ResponseData"
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "$ref": "#/definitions/response.ResponseData"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "dto.UserLoginReq": {
            "type": "object",
            "required": [
                "password",
                "username_or_email"
            ],
            "properties": {
                "password": {
                    "description": "Password 是用户的登录密码。\n`json:\"password\"`: 在 JSON 中此字段名为 \"password\"。\n`binding:\"required\"`: 此字段为必填项。",
                    "type": "string"
                },
                "username_or_email": {
                    "description": "UsernameOrEmail 是用户用于登录的标识，可以是用户名或电子邮件地址。\n`json:\"username_or_email\"`: 在 JSON 中此字段名为 \"username_or_email\"。\n`binding:\"required\"`: 此字段为必填项。",
                    "type": "string"
                }
            }
        },
        "dto.UserLoginRes": {
            "type": "object",
            "properties": {
                "access_token": {
                    "description": "AccessToken 是用户成功登录后获取到的 JWT。\n客户端在后续请求中应将此 Token 包含在 Authorization 请求头中 (通常以 \"Bearer \" 为前缀)。",
                    "type": "string"
                },
                "expires_in": {
                    "description": "ExpiresIn 表示 AccessToken 的过期时间戳 (Unix timestamp in seconds)。\n客户端可以使用此信息来管理 Token 的生命周期，例如在过期前刷新 Token (如果支持刷新令牌)。",
                    "type": "integer"
                },
                "token_type": {
                    "description": "TokenType 表示令牌的类型。对于 JWT，这通常是 \"Bearer\"。",
                    "type": "string"
                },
                "user_id": {
                    "description": "UserID 是成功登录的用户的唯一标识符。\n可选地在登录响应中返回用户ID，便于前端使用。",
                    "type": "integer"
                },
                "username": {
                    "description": "Username 是成功登录的用户的用户名。\n可选地在登录响应中返回用户名，便于前端使用。",
                    "type": "string"
                }
            }
        },
        "dto.UserRegisterReq": {
            "type": "object",
            "required": [
                "email",
                "password",
                "username"
            ],
            "properties": {
                "email": {
                    "description": "Email 是用户的电子邮件地址。\n`json:\"email\"`: 在 JSON 中此字段名为 \"email\"。\n`binding:\"required,email\"`: 此字段为必填项，并且必须符合电子邮件的格式。",
                    "type": "string"
                },
                "password": {
                    "description": "Password 是用户设置的密码。\n`json:\"password\"`: 在 JSON 中此字段名为 \"password\"。\n`binding:\"required,min=8,max=100\"`: 此字段为必填项，长度必须在 8 到 100 个字符之间。",
                    "type": "string",
                    "maxLength": 100,
                    "minLength": 8
                },
                "username": {
                    "description": "Username 是用户期望的登录名。\n`json:\"username\"`: 在 JSON 中此字段名为 \"username\"。\n`binding:\"required,min=3,max=50\"`: 此字段为必填项，长度必须在 3 到 50 个字符之间。",
                    "type": "string",
                    "maxLength": 50,
                    "minLength": 3
                }
            }
        },
        "model.User": {
            "type": "object",
            "properties": {
                "created_at": {
                    "description": "CreatedAt 记录用户创建的时间。\nGORM 会在创建记录时自动填充此字段。",
                    "type": "string"
                },
                "email": {
                    "description": "Email 是用户的电子邮件地址。\n设置了唯一索引 (uniqueIndex) 以确保电子邮件的唯一性。\n允许为空 (default:null)，因为用户可能在注册初期不提供邮箱。",
                    "type": "string"
                },
                "id": {
                    "description": "ID 是用户的主键，类型为 uint，并且在数据库中自增。",
                    "type": "integer"
                },
                "role": {
                    "description": "Role 存储用户的角色，例如 \"role_user\", \"role_admin\"。\ngorm:\"type:varchar(50)\" 指定字段类型。\ngorm:\"default:'role_user'\" 指定数据库中此字段的默认值为 \"role_user\"。\ngorm:\"not null\" 指定字段不能为空。\njson:\"role,omitempty\" 指定 JSON 字段名，并在值为空时从 JSON 输出中省略此字段。",
                    "type": "string"
                },
                "updated_at": {
                    "description": "UpdatedAt 记录用户最后更新的时间。\nGORM 会在创建和更新记录时自动填充此字段。",
                    "type": "string"
                },
                "username": {
                    "description": "Username 是用户的唯一标识名。\n设置了唯一索引 (uniqueIndex) 以确保用户名的唯一性。\n不能为空 (not null)。",
                    "type": "string"
                }
            }
        },
        "response.ResponseData": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "业务状态码",
                    "type": "integer"
                },
                "data": {
                    "description": "响应数据"
                },
                "msg": {
                    "description": "响应消息",
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type \"Bearer\" followed by a space and JWT token. Example: \"Bearer {token}\"",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}